+++ src/main/java/com/writon/admin/domain/dto/request/challenge/CreateChallengeRequestDto.java
@@ -1,5 +1,6 @@
 package com.writon.admin.domain.dto.request.challenge;
 
+import com.writon.admin.domain.entity.lcoal.SpecialQuestion;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -14,7 +15,7 @@ public class CreateChallengeRequestDto {
   private LocalDate endDate;
   private List<LocalDate> dates;
   private List<String> basicQuestions;
-  private List<SpecialQuestionResponse> specialQuestions;
+  private List<SpecialQuestion> specialQuestions;
   private List<String> emailList;
 
 }
