--- src/main/java/com/writon/admin/domain/dto/request/challenge/SpecialQuestionResponse.java
@@ -1,13 +1,0 @@
-package com.writon.admin.domain.dto.request.challenge;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class SpecialQuestionResponse {
-
-  private String keyword;
-  private String[] questions;
-
-}
